feat(consensus): CONS-501a — extract validator_discovery into lib-consensus-net
Migrates `lib-consensus/src/validators/validator_discovery.rs` (783
LOC) into `lib-consensus-net/src/discovery/mod.rs` and updates the
lib-network deep-import to consume from the new location.

## Why partial scope (1 of 4 modules)

The CONS-501 spec lists four modules to move: codec, heartbeat,
validator_protocol, and validator_discovery. Three of them
transitively reference `ConsensusProof`, which contains
`Option<StorageCapacityAttestation>` (lib-storage) and other
`lib-proofs` types — both forbidden in `lib-consensus-core` per
AD-002. Until CONS-201 Scope B resolves that cycle (e.g. by making
the proof fields opaque bytes at the core layer), those modules
have to stay in `lib-consensus`.

`validator_discovery` has zero `crate::` imports — its only
external types come from `lib_crypto`. Clean lift, no cycle.

## Move mechanics

- `lib-consensus-net/src/discovery/mod.rs` now holds the full module
  + its 6 unit tests.
- `lib-consensus/src/validators/validator_discovery.rs` deleted.
- `lib-consensus/src/validators/mod.rs` keeps the path
  `crate::validators::validator_discovery::*` working via a re-export
  module `pub mod validator_discovery { pub use lib_consensus_net::
  discovery::*; }` — every internal lib-consensus caller compiles
  unchanged.
- `lib-network/src/validator_discovery_transport.rs` now imports
  directly from `lib_consensus_net::discovery::*` (acceptance:
  `grep "lib_consensus::validators::validator_discovery" --include
  "*.rs"` outside `lib-consensus/src` returns zero hits — only
  doc comments mentioning the path remain).

## Cycle break

The original `lib-consensus-net/Cargo.toml` listed `lib-network` as a
dep (speculative for the future module moves). Adding
`lib-consensus-net` to `lib-consensus` would have created the cycle
`lib-blockchain → lib-network → lib-consensus-net → lib-consensus →
lib-blockchain`. Since the discovery module needs nothing from
`lib-network`, the dep is dropped from `lib-consensus-net` for now.
It will be re-added when codec/heartbeat/validator_protocol move —
at which point CONS-507 needs to retire `lib-consensus`'s
`lib-network` use first to keep the graph acyclic.

## Validation

- `cargo build --workspace` — clean (2m 09s warm).
- `cargo test -p lib-consensus --lib` — **181/181** (was 187 before
  the 6 discovery tests moved with the file).
- `cargo test -p lib-consensus-net --lib` — **6/6** pass (the 6
  tests that moved).
- `cargo test -p lib-consensus-core --lib` — **41/41**.
- `cargo test -p lib-consensus-runtime --lib` — **15/15**.
- Acceptance grep: zero `lib_consensus::validators::validator_discovery`
  deep-imports in code outside `lib-consensus/src`.

## CONS-501 remainder tracked separately

codec / heartbeat / validator_protocol moves are blocked on
CONS-201 Scope B. Issue #2351 stays open until those land.

Refs #2351
