+++ lib-consensus-net/Cargo.toml
@@ -11,10 +11,17 @@ description = "Wire format + transport adapters for lib-consensus-core."
 # validators/validator_protocol.rs,validators/validator_discovery.rs}.
 [dependencies]
 lib-consensus-core = { path = "../lib-consensus-core" }
-lib-network        = { path = "../lib-network" }
+# CONS-501: `lib-network` is intentionally NOT a dep here. The discovery
+# module migrated in this PR has zero `lib-network` references; the
+# remaining stub modules (codec, heartbeat, validator_protocol) will
+# need it when CONS-201 Scope B unblocks their move, at which point
+# the dep is re-added — but only after the lib-blockchain → lib-network
+# → lib-consensus-net → lib-consensus → lib-blockchain cycle is broken
+# by retiring `lib-consensus`'s `lib-network` use (CONS-507).
 lib-crypto         = { path = "../lib-crypto" }
 lib-identity       = { path = "../lib-identity" }
 lib-types          = { path = "../lib-types" }
+anyhow             = "1.0"
 serde              = { version = "1.0", features = ["derive"] }
 thiserror          = "1.0"
 tracing            = "0.1"
